requirements.txt
anyio==2.0.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jplephem==2.15
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
kiwisolver==1.3.1
lunardate==0.2.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5.1
notebook==6.2.0
numpy==1.19.4
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pyCalverter==1.6.1
pycparser==2.20
Pygments==2.7.4
pyluach==1.2.1
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.5.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pyzmq==21.0.1
requests==2.25.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
sgp4==2.14
six==1.15.0
sklearn==0.0
skyfield==1.35
skyfield-data==3.0.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.7.1
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
workalendar==14.1.0
